Rename local module
